Switching from antd to Material-ui and starting move towards a duck-type project structure
Reason for changing to Material-UI: although I have experimented with other libraries (the previous one was Ant Design), I am used to Material-UI from my job and I find it better documented, designed and supported.

Login module has been refactored to use duck project structure (https://hackernoon.com/my-journey-toward-a-maintainable-project-structure-for-react-redux-b05dfd999b5).
The reasons for this is that I find that the previous structure couples too much actions and reducers with the components that use them.
So, let's experiment with this structure to see if I find it neater.

Also, I made a general dependency upgrade.